Splitta opp login slik at routes og logikk er separert
